+++ ccvm_simulators/solvers/mf_solver.py
@@ -148,6 +148,8 @@ class MFSolver(CCVMSolver):
         g,
         S,
         fs,
+        lower_limit=0,
+        upper_limit=1,
     ):
         """We treat the SDE that simulates the CIM of NTT as drift
         calculation.
@@ -160,7 +162,9 @@ class MFSolver(CCVMSolver):
             j (float): The measurement strength
             g (float): The nonlinearity coefficient
             S (float): The enforced saturation value
-            fs (float): The coefficient of the feedback term.
+            fs (float): The coefficient of the feedback term
+            lower_limit (float): The lower bound of the box constraint
+            upper_limit (float): The upper bound of the box constraint
 
         Returns:
             tuple: The gradients of the mean-field amplitudes and the variance.
@@ -168,12 +172,21 @@ class MFSolver(CCVMSolver):
         mu_pow = torch.pow(mu, 2)
 
         mu_term1 = (-(1 + j) + pump - g**2 * mu_pow) * mu
+
         mu_term2_1 = (
             -(1 / 4)
-            * (torch.einsum("bi,ij -> bj", mu_tilde / S + 1, self.q_matrix))
+            * (
+                torch.einsum(
+                    "bi,ij -> bj",
+                    mu_tilde * (upper_limit - lower_limit) / S
+                    + (upper_limit + lower_limit),
+                    self.q_matrix,
+                )
+            )
+            * (upper_limit - lower_limit)
             / S
         )
-        mu_term2_2 = -self.v_vector / S / 2
+        mu_term2_2 = -self.v_vector * (upper_limit - lower_limit) / (2 * S)
 
         sigma_term1 = 2 * (-(1 + j) + pump - 3 * g**2 * mu_pow) * sigma
         sigma_term2 = -2 * j * (sigma - 0.5).pow(2)
@@ -184,7 +197,7 @@ class MFSolver(CCVMSolver):
 
         return drift_mu, drift_sigma
 
-    def _calculate_grads_boxqp(self, mu_tilde, S, fs):
+    def _calculate_grads_boxqp(self, mu_tilde, S, fs, lower_limit=0, upper_limit=1):
         """We treat the SDE that simulates the CIM of NTT as gradient
         calculation. Original SDE considers only quadratic part of the objective
         function. Therefore, we need to modify and add linear part of the QP to
@@ -193,33 +206,49 @@ class MFSolver(CCVMSolver):
         Args:
             mu_tilde (torch.Tensor): Mean-field measured amplitudes
             S (float): The enforced saturation value
-            fs (float): The coefficient of the feedback term.
+            fs (float): The coefficient of the feedback term
+            lower_limit (float): The lower bound of the box constraints
+            upper_limit (float): The upper bound of the box constraints
 
         Returns:
             tensor: The gradients of the mean-field amplitude.
         """
         mu_term2_1 = (
             -(1 / 4)
-            * (torch.einsum("bi,ij -> bj", mu_tilde / S + 1, self.q_matrix))
+            * (
+                torch.einsum(
+                    "bi,ij -> bj",
+                    mu_tilde * (upper_limit - lower_limit) / S
+                    + (upper_limit + lower_limit),
+                    self.q_matrix,
+                )
+            )
+            * (upper_limit - lower_limit)
             / S
         )
-        mu_term2_2 = -self.v_vector / S / 2
+        mu_term2_2 = -self.v_vector * (upper_limit - lower_limit) / (2 * S)
 
         grads_mu = fs * (mu_term2_1 + mu_term2_2)
 
         return grads_mu
 
-    def _change_variables_boxqp(self, problem_variables, S=1):
+    def _change_variables_boxqp(
+        self, problem_variables, lower_limit=0, upper_limit=1, S=1
+    ):
         """Perform a change of variables to enforce the box constraints.
 
         Args:
             problem_variables (torch.Tensor): The variables to change.
+            lower_limit (float): The lower bound of the box constraints. Defaults to 0.
+            upper_limit (float): The upper bound of the box constraints. Defaults to 1.
             S (float or torch.tensor): The enforced saturation value. Defaults to 1
 
         Returns:
             torch.Tensor: The changed variables.
         """
-        return 0.5 * (problem_variables / S + 1)
+        return 0.5 * problem_variables / S * (upper_limit - lower_limit) + 0.5 * (
+            upper_limit + lower_limit
+        )
 
     def _fit_to_constraints_boxqp(self, mu_tilde, lower_clamp, upper_clamp):
         """Clamps the values of mu_tilde to be within the box constraints
@@ -538,6 +567,8 @@ class MFSolver(CCVMSolver):
                 g,
                 S,
                 feedback_scale,
+                self.solution_bounds[0],
+                self.solution_bounds[1],
             )
 
             # mu_term3-> mu_diffusion
@@ -679,6 +710,8 @@ class MFSolver(CCVMSolver):
                 mu_tilde_c,
                 S,
                 feedback_scale,
+                self.solution_bounds[0],
+                self.solution_bounds[1],
             )
 
             # Update biased first moment estimate
@@ -782,6 +815,7 @@ class MFSolver(CCVMSolver):
         problem_size = instance.problem_size
         self.q_matrix = instance.q_matrix
         self.v_vector = instance.v_vector
+        self.solution_bounds = instance.solution_bounds
 
         # Get solver setup variables
         batch_size = self.batch_size
@@ -897,12 +931,18 @@ class MFSolver(CCVMSolver):
             )
 
             problem_variables = post_processor_object.postprocess(
-                self.change_variables(mu_tilde, S), self.q_matrix, self.v_vector
+                self.change_variables(
+                    mu_tilde, self.solution_bounds[0], self.solution_bounds[1], S
+                ),
+                self.q_matrix,
+                self.v_vector,
             )
             # Post-processing time for solving an instance once
             pp_time = post_processor_object.pp_time / batch_size
         else:
-            problem_variables = self.change_variables(mu_tilde, S)
+            problem_variables = self.change_variables(
+                mu_tilde, self.solution_bounds[0], self.solution_bounds[1], S
+            )
             pp_time = 0.0
 
         objval = instance.compute_energy(problem_variables)